Updated luminance functions to use integer values instead of bytes. Closes

+++ src/main/java/org/spongepowered/api/world/Location.java
@@ -492,7 +492,7 @@ public final class Location implements DataHolder {
      *
      * @return A light level, nominally between 0 and 15, inclusive
      */
-    public byte getLuminance() {
+    public int getLuminance() {
         return getExtent().getLuminance(getBlockPosition());
     }
 
@@ -504,7 +504,7 @@ public final class Location implements DataHolder {
      *
      * @return A light level, nominally between 0 and 15, inclusive
      */
-    public byte getLuminanceFromSky() {
+    public int getLuminanceFromSky() {
         return getExtent().getLuminanceFromSky(getBlockPosition());
     }
 
@@ -516,7 +516,7 @@ public final class Location implements DataHolder {
      *
      * @return A light level, nominally between 0 and 15, inclusive
      */
-    public byte getLuminanceFromGround() {
+    public int getLuminanceFromGround() {
         return getExtent().getLuminanceFromGround(getBlockPosition());
     }
